feat: add assignee autocomplete from existing tickets
