+++ src/webview/App.tsx
@@ -34,6 +34,10 @@ function App(): React.JSX.Element {
     frontmatter: FeatureFrontmatter
     contentVersion: number
   } | null>(null)
+  const editingFeatureRef = useRef(editingFeature)
+  useEffect(() => {
+    editingFeatureRef.current = editingFeature
+  }, [editingFeature])
 
   // Undo delete stack
   const [pendingDeletes, setPendingDeletes] = useState<{ id: string; feature: Feature }[]>([])
@@ -199,7 +203,7 @@ function App(): React.JSX.Element {
           setColumns(message.columns)
           setCollapsedColumns(message.collapsedColumns ?? [])
           if (message.settings) {
-            if (message.settings.markdownEditorMode && editingFeature) {
+            if (message.settings.markdownEditorMode && editingFeatureRef.current) {
               setEditingFeature(null)
             }
             setCardSettings(message.settings)
